backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
openai>=1.60.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
httpx>=0.28.0
mcp>=1.3.0
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langchain-mcp-adapters>=0.1.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-cloud-texttospeech>=2.16.0
yfinance>=0.2.40
langmem>=0.0.30
aiosqlite>=0.22.0
firebase-admin>=6.5.0
